requirements.txt
python-dotenv==0.21.1
requests==2.32.3
pydantic==2.9.2
rich==13.9.4
feedparser==6.0.11

